+++ src/libft/utils/clamp.c
@@ -6,13 +6,13 @@
 /*   By: omimouni <omimouni@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/01 15:49:48 by omimouni          #+#    #+#             */
-/*   Updated: 2021/02/09 09:00:15 by omimouni         ###   ########.fr       */
+/*   Updated: 2021/02/16 09:27:12 by omimouni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_clamp_int(int i, int min, int max)
+long	ft_clamp_int(long i, long min, long max)
 {
 	if (i < min)
 		return (min);
